airflow_local/requirements.txt
requests
pandas
sqlalchemy
psycopg2-binary
pyyaml